+++ tests/Console/Command/Code/execute.php
@@ -1,0 +1,3 @@
+<?php
+
+$GLOBALS['current_user']->id;