+++ tests/Console/Command/Code/sugar_exception.php
@@ -1,0 +1,3 @@
+<?php
+
+throw new \Inet\SugarCRM\Exception\SugarException('error from sugar');