+++ CHANGELOG.md
@@ -3,12 +3,13 @@ Changelog
 
 (unreleased)
 ----
-* Optional `--url` parameter in `install:run` command if url is set in configuration file.
-* Add `--email` option for `user:update` command to set email address of a user.
-* Add `--ask-password` option for `user:update` command to ask for password
+* New command `code:execute:file` to run a php script from SugarCRM context.
 * Option `sugarcrm.path` is now relative to the configuration file instead of current directory.
 * Search for `.sugarclirc` files in all parent folders of the current directory. Configuration will
   be overriden by the children.
+* Optional `--url` parameter in `install:run` command if url is set in configuration file.
+* Add `--email` option for `user:update` command to set email address of a user.
+* Add `--ask-password` option for `user:update` command to ask for password
 
 1.9.1
 ----